requirements.txt
typing-extensions==4.6.2
sentencepiece==0.1.99
numpy==1.24.3
pandas==1.5.3
pytorch_lightning==1.9.5
tqdm==4.65.0
transformers==4.31.0
openai==0.27.7
Requests==2.31.0
datasets>=1.17.0
accelerate
peft>=0.3.0
tyro>=0.5.7
trl